- Added Sunrise, Moonshine, and Bloomlight Paper Lanterns

+++ src/main/java/net/sydokiddo/lottablocks/block/custom_blocks/LavaLampBlock.java
@@ -27,7 +27,7 @@ public class LavaLampBlock extends PillarBlock {
             }
 
             if (random.nextInt(200) == 0) {
-                world.playSound((double)pos.getX(), (double)pos.getY(), (double)pos.getZ(), SoundEvents.BLOCK_LAVA_AMBIENT, SoundCategory.BLOCKS, 0.2F + random.nextFloat() * 0.2F, 0.9F + random.nextFloat() * 0.15F, false);
+                world.playSound(pos.getX(), pos.getY(), pos.getZ(), SoundEvents.BLOCK_LAVA_AMBIENT, SoundCategory.BLOCKS, 0.2F + random.nextFloat() * 0.2F, 0.9F + random.nextFloat() * 0.15F, false);
             }
         }
     }
